+++ tests/testthat/test-cpm.R
@@ -5,7 +5,7 @@ test_that("Default threshold method works", {
   result <- cpm(conmat, behav)
   expect_s3_class(result, "cpm")
   expect_snapshot_value(result$pred, style = "json2")
-  expect_snapshot_value(result$edges, style = "json2")
+  expect_null(result$edges)
   expect_snapshot(result)
 })
 
@@ -16,7 +16,7 @@ test_that("`kfolds` works", {
   result <- cpm(conmat, behav, kfolds = 5)
   expect_s3_class(result, "cpm")
   expect_snapshot_value(result$pred, style = "json2")
-  expect_snapshot_value(result$edges, style = "json2")
+  expect_null(result$edges)
   expect_snapshot(result)
 })
 
@@ -27,7 +27,8 @@ test_that("Alternative threshold method works", {
   result <- cpm(conmat, behav, thresh_method = "sparsity")
   expect_s3_class(result, "cpm")
   expect_snapshot_value(result$pred, style = "json2")
-  expect_snapshot_value(result$edges, style = "json2")
+  expect_null(result$edges)
+  expect_snapshot(result)
 })
 
 test_that("Different threshold levels works", {
@@ -37,7 +38,8 @@ test_that("Different threshold levels works", {
   result <- cpm(conmat, behav, thresh_level = 0.1)
   expect_s3_class(result, "cpm")
   expect_snapshot_value(result$pred, style = "json2")
-  expect_snapshot_value(result$edges, style = "json2")
+  expect_null(result$edges)
+  expect_snapshot(result)
 })
 
 test_that("Works with confounds", {
@@ -48,7 +50,8 @@ test_that("Works with confounds", {
   result <- cpm(conmat, behav, confounds = confounds)
   expect_s3_class(result, "cpm")
   expect_snapshot_value(result$pred, style = "json2", tolerance = 1e-6)
-  expect_snapshot_value(result$edges, style = "json2")
+  expect_null(result$edges)
+  expect_snapshot(result)
 })
 
 test_that("Keep names of behavior", {
@@ -69,4 +72,43 @@ test_that("`return_edges` argument works", {
   expect_snapshot_value(result$edges, style = "json2")
   result <- cpm(conmat, behav, return_edges = "sum")
   expect_snapshot_value(result$edges, style = "json2")
+})
+
+test_that("Support row/column matrix input of `behav` and `confounds`", {
+  withr::local_seed(123)
+  conmat <- matrix(rnorm(100), ncol = 10)
+  behav <- rnorm(10)
+  result <- cpm(conmat, behav)
+  key_fields <- c("real", "pred", "edges")
+  expect_identical(
+    cpm(conmat, matrix(behav, ncol = 1))[key_fields],
+    result[key_fields]
+  )
+  expect_identical(
+    cpm(conmat, matrix(behav, nrow = 1))[key_fields],
+    result[key_fields]
+  )
+  confounds <- matrix(rnorm(10), ncol = 1)
+  result <- cpm(conmat, behav, confounds = confounds)
+  expect_identical(
+    cpm(conmat, behav, confounds = drop(confounds))[key_fields],
+    result[key_fields]
+  )
+})
+
+test_that("Throw informative error if data checking not pass", {
+  withr::local_seed(123)
+  conmat <- matrix(rnorm(100), ncol = 10)
+  expect_error(
+    cpm(conmat, matrix(rnorm(20), ncol = 2)),
+    "Behavior data must be a numeric vector."
+  )
+  expect_error(
+    cpm(conmat, rnorm(20)),
+    "Case numbers of `conmat` and `behav` must match."
+  )
+  expect_error(
+    cpm(conmat, rnorm(10), confounds = matrix(rnorm(20), ncol = 1)),
+    "Case numbers of `confounds` and `behav` must match."
+  )
 })